implementovany USR2

+++ proj01.c
@@ -8,18 +8,15 @@
 #include <sys/types.h>
 #include <stdio.h>
 
-volatile sig_atomic_t signaled;
+volatile sig_atomic_t signaled = 0;
 volatile sig_atomic_t reset = 0;
 
 void sig_usr1(int sig){
-
     signaled = 1;
 }
 
 void sig_usr2(int sig){
-
     reset = 1;
-    printf("%d:id %d\n",sig,getpid());
 }
 
 int main(void)
@@ -27,15 +24,11 @@ int main(void)
     pid_t pid,ppid;
     struct sigaction sigact1;
     struct sigaction sigact2;
-    sigset_t empty_set;
     sigset_t sigusr2_set;
     int char_num = ABC_START;
     int c;
 
     ppid = getpid();
-
-    sigemptyset(&empty_set);
-
     sigemptyset(&sigusr2_set);
     sigaddset(&sigusr2_set,SIGUSR2);
 
@@ -61,24 +54,25 @@ int main(void)
     /* parrent process */
     if(pid > 0){
         while (1){
-
-
-            if(char_num == ABC_END) char_num = ABC_START;
-            printf("Parent (%d): '%c'\n",ppid,char_num++);
-            kill(pid, SIGUSR1);
-
-            /* pockaj na singal USR1 a ignoruj USR2 */
+            if(reset) {
+                sigprocmask(SIG_BLOCK, &sigusr2_set,NULL);
+                char_num = ABC_START;
+                reset = 0;
+                signaled = 1;/* pokracuj na getchar() */
+                sigprocmask(SIG_UNBLOCK, &sigusr2_set,NULL);
+            }else{
+                if(char_num == ABC_END) char_num = ABC_START;
+                printf("Parent (%d): '%c'\n",ppid,char_num++);
+                kill(pid, SIGUSR1);
+                printf("Press enter...");
+            }
+            /* pockaj na singal USR1 a blokuj USR2 */
             if(!signaled) sigsuspend(&sigusr2_set);
             signaled = 0;
 
-            /* len enter */
-            printf("Press enter...");
+            /* treba pockat na enter */
             c = getchar();
             while (c != '\n' && c != EOF) c=getchar();
-
-            /* DOPRACUJ USR2 */
-            /*sigprocmask(SIG_BLOCK, &sigusr2_set,NULL);*/
-            /*sigprocmask(SIG_UNBLOCK, &sigusr2_set,NULL);*/
         }
     }
 
@@ -87,19 +81,20 @@ int main(void)
     if(pid == 0) {
         while (1){
 
-            /* pockaj na singal USR1 a ignoruj USR2 */
+            /* pockaj na singal USR1 a blokuj USR2 */
             if(!signaled) sigsuspend(&sigusr2_set);
-
             signaled = 0;
+
+            if(reset){
+                sigprocmask(SIG_BLOCK, &sigusr2_set,NULL);
+                reset = 0;
+                char_num = ABC_START;
+                sigprocmask(SIG_UNBLOCK, &sigusr2_set,NULL);
+            }
+
             if(char_num == ABC_END) char_num = ABC_START;
             printf("Child (%d): '%c'\n",getpid(),char_num++);
             kill(ppid, SIGUSR1);
-
-
-
-            /* DOPRACUJ USR2 */
-            /*sigprocmask(SIG_BLOCK, &sigusr2_set,NULL);*/
-            /*sigprocmask(SIG_UNBLOCK, &sigusr2_set,NULL);*/
         }
     }
